+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v60.2.0...main)
+[Full changelog](https://github.com/mozilla/glean/compare/v60.3.0...main)
+
+# v60.3.0 (2024-05-31)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v60.2.0...v60.3.0)
+
+* Android
+  * Allow configuring `delayPingLifetimeIo` in Kotlin and auto-flush this data after 1000 writes.
+    It is also auto-flushed on background. ([#2851](https://github.com/mozilla/glean/pull/2851))
 
 # v60.2.0 (2024-05-23)
 
@@ -8,6 +16,15 @@
 
 * Rust
   * Accept a ping schedule map on initialize ([#2839](https://github.com/mozilla/glean/pull/2839))
+
+# v60.1.1 (2024-05-31)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v60.1.0...v60.1.1)
+
+* Android
+  * Allow configuring `delayPingLifetimeIo` in Kotlin and auto-flush this data after 1000 writes.
+    It is also auto-flushed on background. ([#2851](https://github.com/mozilla/glean/pull/2851))
+    (Backported changes)
 
 # v60.1.0 (2024-05-06)